Split log messages into log, warn and error
We would like to log only warning and errors and omit other info messages which will be logged if logging is enabled. Therefore I added the functions LDAP.warn and LDAP.error which pass their message to the LDAP.log function per default. They both can now be overwritten to handle the errors and warnings separately, if desired. I scanned all usages of the LDAP.log function and replaced them by LDAP.warn or LDAP.error depending on the purpose.

Additionally I added a paragraph in the README.md to describe the new functions.